mm color + bug fixes and other things

+++ common/src/main/java/net/refractionapi/refraction/feature/camerashake/CameraShakeHandler.java
@@ -10,11 +10,11 @@ import net.refractionapi.refraction.networking.S2C.InvokeCameraShakeS2CPacket;
 
 public class CameraShakeHandler {
 
-    public static void invokeCameraShakeToPlayer(int duration, int intensity, ServerPlayer serverPlayer) {
+    public static void invokeCameraShakeToPlayer(int duration, float intensity, ServerPlayer serverPlayer) {
         RefractionMessages.sendToPlayer(new InvokeCameraShakeS2CPacket(duration, intensity), serverPlayer);
     }
 
-    public static void invokeCameraShakeToPlayersWithinRange(int duration, int intensity, int range, BlockPos pos, ServerLevel level) {
+    public static void invokeCameraShakeToPlayersWithinRange(int duration, float intensity, int range, BlockPos pos, ServerLevel level) {
         AABB aabb = new AABB(pos.getX() - range, pos.getY() - range, pos.getZ() - range, pos.getX() + range, pos.getY() + range, pos.getZ() + range);
         for (Entity entity : level.getEntities(null, aabb)) {
             if (entity instanceof ServerPlayer player) {
@@ -23,7 +23,7 @@ public class CameraShakeHandler {
         }
     }
 
-    public static void invokeCameraShakeToPlayersWithinRange(int duration, int intensity, int range, ServerPlayer serverPlayer) {
+    public static void invokeCameraShakeToPlayersWithinRange(int duration, float intensity, int range, ServerPlayer serverPlayer) {
         AABB aabb = new AABB(serverPlayer.getX() - range, serverPlayer.getY() - range, serverPlayer.getZ() - range, serverPlayer.getX() + range, serverPlayer.getY() + range, serverPlayer.getZ() + range);
         for (Entity entity : serverPlayer.level().getEntities(null, aabb)) {
             if (entity instanceof ServerPlayer player) {
