+++ common/src/main/java/net/refractionapi/refraction/mixininterfaces/ICameraMixin.java
@@ -2,6 +2,6 @@ package net.refractionapi.refraction.mixininterfaces;
 
 public interface ICameraMixin {
 
-    void startCameraShake(int durationInTicks, int intensity);
+    void startCameraShake(int durationInTicks, float intensity);
 
 }
